Move thirdParty to common/resources
