+++ services/user/CommonSys/common/.gitignore
@@ -1,1 +1,1 @@
-thirdParty/*
+resources/thirdParty/*
