+++ services/user/ExploreSys/ui/vite.config.ts
@@ -11,7 +11,7 @@ const psibase = (appletContract: string, isServing?: boolean) => {
         {
             find: "/common/iframeResizer.contentWindow.js",
             replacement: path.resolve(
-                "../../CommonSys/common/thirdParty/src/iframeResizer.contentWindow.js"
+                "../../CommonSys/common/resources/thirdParty/src/iframeResizer.contentWindow.js"
             ),
         },
         {
